app/requirements.txt
flask
gunicorn
psycopg2-binary
mysql-connector-python
redis
